explicitly define order of creation
Terraform resources are usually created in a dependency-resolved manner. However, in this case if we have to set the desired count as 2, then the ALB is throwing a 503 as target groups does not getting registered in time. 

I had to explicitly define dependencies using depends_on in your aws_autoscaling_attachment resource.